Moved all the example files under examples/Modal/
